Hi,

With the latest changes to the sorting algorithm, the sorting order was no longer as intuitive as it could be.

For example, in the current code, sorting by 'year' with years: [914, 1066, 1929, 2019] results in [1066, 1929, 2019, 914].

Similarly, the current sorting algorithm brings 'None' values to the top of the sort, rather than the end, where IMO they more intuitively belong.

Unlike my first attempt at this sorting algorithm, this doesn't sort negative numbers properly, which means it can use the (faster) isnumeric() function rather than the slower try/except block.

Is this OK to go in? 

Jackson