@alejandrogallo 2020-01-05
Hi,
thank you for this, it sounds reasonable. I'll take a look at it and merge it when I'm done
implementing typing for papis, which hopefully will be this weekend.
Cheers,
Ale