@j-c-w 2020-01-11
Since performance was a concern before, I thought I would go through and check the performance of these two approaches to sorting, for 100,000 items to sort (i.e. far more than one would reasonably expect papis to handle) is:

Current approach: 0.085 to 0.090 seconds
Approach involving sorting numbers correctly: 0.160 to 0.166 seconds

Script for the comparison is here:
[sorting_speed_comparison.py.txt](https://github.com/papis/papis/files/4049173/sorting_speed_comparison.py.txt) (rename to .py then run `python3 sorting_speed_comparison.py`)